Poprawka ze sprawdzaniem uprawnien

+++ grails-app/domain/pl/fork/place/other/PricedElement.groovy
@@ -9,6 +9,6 @@ class PricedElement {
 
     String name
     String description
-    int price
+    double price
 
 }
